requirements.txt
streamlit
langchain
langchain-openai
faiss-cpu
google-generativeai
pandas
numpy
pathlib
openai 